Degisiklikler

+++ Controllers/BasketController.cs
@@ -1,4 +1,6 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using OceanAPI.NET6.Dtos;
 using OceanAPI.NET6.Models;
@@ -10,48 +12,59 @@ namespace OceanAPI.NET6.Controllers
     [ApiController]
     public class BasketController : ControllerBase
     {
-        private readonly IBasketService _basketService;
         private readonly IBasketTransactionService _basketTransactionService;
         private readonly IMapper _mapper;
         public BasketController(IBasketService basketService, IBasketTransactionService basketTransactionService, IMapper mapper)
         {
-            _basketService = basketService;
             _mapper = mapper;
             _basketTransactionService = basketTransactionService;
         }
 
         [HttpGet("{basketId}")]
-        public ActionResult GetBasket(int basketId)
+        public async Task<ActionResult> GetBasket(int basketId)
         {
-            var basket = _basketService.GetBasket(basketId);
+            if (!Extensions.IsCurrentUser(basketId, User))
+                return Forbid();
+            var basket = await _basketTransactionService.GetBasket(basketId);
             if (basket == null)
                 return NotFound();
             var basketDto = _mapper.Map<Basket,BasketDto>(basket);
             return Ok(basketDto);
         }
 
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         [HttpPost("{id}/products")]
-        public ActionResult AddProduct(int id, BasketProductCreateDto basketProductCreateDto)
+        public async Task<ActionResult> AddProduct(int id, BasketProductCreateDto basketProductCreateDto)
         {
-            var response = _basketTransactionService.AddProduct(id, basketProductCreateDto);
+            if (!Extensions.IsCurrentUser(id, User))
+                return Forbid();
+            var response = await _basketTransactionService.AddProduct(id, basketProductCreateDto);
             if (response == null)
                 return BadRequest();
             var basketDto = _mapper.Map<Basket, BasketDto>(response);
             return Ok(basketDto);
         }
+
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         [HttpPut("{id}/products")]
-        public ActionResult ChangeProduct(int id, BasketProductCreateDto basketProductCreateDto)
+        public async Task<ActionResult> ChangeProduct(int id, BasketProductCreateDto basketProductCreateDto)
         {
-            var response = _basketTransactionService.ChangeProduct(id, basketProductCreateDto);
+            if (!Extensions.IsCurrentUser(id, User))
+                return Forbid();
+            var response = await _basketTransactionService.ChangeProduct(id, basketProductCreateDto);
             if (response == null)
                 return BadRequest();
             var basketDto = _mapper.Map<Basket, BasketDto>(response);
             return Ok(basketDto);
         }
+
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         [HttpDelete("{id}/products/{productId}")]
-        public ActionResult RemoveProduct(int id, int productId)
+        public async Task<ActionResult> RemoveProduct(int id, int productId)
         {
-            var response = _basketTransactionService.RemoveProduct(id, productId);
+            if (!Extensions.IsCurrentUser(id, User))
+                return Forbid();
+            var response = await _basketTransactionService.RemoveProduct(id, productId);
             if (response == null)
                 return BadRequest();
             var basketDto = _mapper.Map<Basket, BasketDto>(response);
